+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Nullable union property with null value cannot be mapped [PR#200](https://github.com/JsonMapper/JsonMapper/pull/200)
 
 ## [2.25.0] - 2025-04-29
 ### Fixed