+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.25.1] - 2025-05-26
+### Fixed
+- Nullable union property with null value cannot be mapped [PR#200](https://github.com/JsonMapper/JsonMapper/pull/200)
+
 ## [2.25.0] - 2025-04-29
 ### Fixed
 - Replace docblock type parsing with reflection docblock library. [PR#199](https://github.com/JsonMapper/JsonMapper/pull/199)